p2p/enode: move ID tests to node_test.go
