+++ p2p/enode/urlv4_test.go
@@ -17,14 +17,11 @@
 package enode
 
 import (
-	"bytes"
 	"crypto/ecdsa"
-	"math/big"
 	"net"
 	"reflect"
 	"strings"
 	"testing"
-	"testing/quick"
 )
 
 var parseNodeTests = []struct {
@@ -159,85 +156,5 @@ func TestNodeString(t *testing.T) {
 				t.Errorf("test %d: Node.String() mismatch:\ngot:  %s\nwant: %s", i, str, test.rawurl)
 			}
 		}
-	}
-}
-
-func TestHexID(t *testing.T) {
-	ref := ID{0, 0, 0, 0, 0, 0, 0, 128, 106, 217, 182, 31, 165, 174, 1, 67, 7, 235, 220, 150, 66, 83, 173, 205, 159, 44, 10, 57, 42, 161, 26, 188}
-	id1 := HexID("0x00000000000000806ad9b61fa5ae014307ebdc964253adcd9f2c0a392aa11abc")
-	id2 := HexID("00000000000000806ad9b61fa5ae014307ebdc964253adcd9f2c0a392aa11abc")
-
-	if id1 != ref {
-		t.Errorf("wrong id1\ngot  %v\nwant %v", id1[:], ref[:])
-	}
-	if id2 != ref {
-		t.Errorf("wrong id2\ngot  %v\nwant %v", id2[:], ref[:])
-	}
-}
-
-func TestID_textEncoding(t *testing.T) {
-	ref := ID{
-		0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x08, 0x09, 0x10,
-		0x11, 0x12, 0x13, 0x14, 0x15, 0x16, 0x17, 0x18, 0x19, 0x20,
-		0x21, 0x22, 0x23, 0x24, 0x25, 0x26, 0x27, 0x28, 0x29, 0x30,
-		0x31, 0x32,
-	}
-	hex := "0102030405060708091011121314151617181920212223242526272829303132"
-
-	text, err := ref.MarshalText()
-	if err != nil {
-		t.Fatal(err)
-	}
-	if !bytes.Equal(text, []byte(hex)) {
-		t.Fatalf("text encoding did not match\nexpected: %s\ngot:      %s", hex, text)
-	}
-
-	id := new(ID)
-	if err := id.UnmarshalText(text); err != nil {
-		t.Fatal(err)
-	}
-	if *id != ref {
-		t.Fatalf("text decoding did not match\nexpected: %s\ngot:      %s", ref, id)
-	}
-}
-
-func TestNodeID_distcmp(t *testing.T) {
-	distcmpBig := func(target, a, b ID) int {
-		tbig := new(big.Int).SetBytes(target[:])
-		abig := new(big.Int).SetBytes(a[:])
-		bbig := new(big.Int).SetBytes(b[:])
-		return new(big.Int).Xor(tbig, abig).Cmp(new(big.Int).Xor(tbig, bbig))
-	}
-	if err := quick.CheckEqual(DistCmp, distcmpBig, nil); err != nil {
-		t.Error(err)
-	}
-}
-
-// The random tests is likely to miss the case where a and b are equal,
-// this test checks it explicitly.
-func TestNodeID_distcmpEqual(t *testing.T) {
-	base := ID{0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15}
-	x := ID{15, 14, 13, 12, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 1, 0}
-	if DistCmp(base, x, x) != 0 {
-		t.Errorf("DistCmp(base, x, x) != 0")
-	}
-}
-
-func TestNodeID_logdist(t *testing.T) {
-	logdistBig := func(a, b ID) int {
-		abig, bbig := new(big.Int).SetBytes(a[:]), new(big.Int).SetBytes(b[:])
-		return new(big.Int).Xor(abig, bbig).BitLen()
-	}
-	if err := quick.CheckEqual(LogDist, logdistBig, nil); err != nil {
-		t.Error(err)
-	}
-}
-
-// The random tests is likely to miss the case where a and b are equal,
-// this test checks it explicitly.
-func TestNodeID_logdistEqual(t *testing.T) {
-	x := ID{0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15}
-	if LogDist(x, x) != 0 {
-		t.Errorf("LogDist(x, x) != 0")
 	}
 }
